reorganization of the structure
