+++ models/index.ts
@@ -1,8 +1,8 @@
 export * from './user.model'
 export * from './role.model'
-export * from './session.model'
-export * from './space.model'
-export * from './animal.model'
-export * from './animalGroup.model'
-export * from './healthBooklet.model'
-export * from './maintenanceBooklet.model'
+export * from './space/session.model'
+export * from './space/space.model'
+export * from './space/animal.model'
+export * from './space/animalGroup.model'
+export * from './space/healthBooklet.model'
+export * from './space/maintenanceBooklet.model'
