+++ models/space/animal.model.ts
