+++ models/space/animalGroup.model.ts
